+++ Hammer.MODEL/Models/HeadParameters.cs
@@ -5,27 +5,27 @@
         /// <summary>
         /// Переменная высоты бойка
         /// </summary>
-        public int Height { get; set; }
+        public double Height { get; set; }
 
         /// <summary>
         /// Переменная длины оголовья
         /// </summary>
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         /// <summary>
         /// Переманная ширины оголовья
         /// </summary>
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         /// <summary>
         /// Переменная ширины наконечника
         /// </summary>
-        public int TipWidth { get; set; }
+        public double TipWidth { get; set; }
 
         /// <summary>
         /// Переменная длины носка
         /// </summary>
-        public int ToeLength { get; set; }
+        public double ToeLength { get; set; }
 
         public HeadParameters() { }
     }
